# docs/swagger.yaml
definitions:
...
  main.Paginate-main_Reservation:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/main.Reservation'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_page:
        type: integer
    type: object
...
  main.Reservation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      reservation_items:
        description: relation
        items:
          $ref: '#/definitions/main.ReservationItem'
        type: array
      status:
        $ref: '#/definitions/main.ReservationStatus'
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  main.ReservationFilter:
    properties:
      ids:
        items:
          type: integer
        type: array
      statuses:
        items:
          type: string
        type: array
      user_ids:
        items:
          type: integer
        type: array
      with_items:
        type: boolean
    type: object
  main.ReservationItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      movie:
        description: relation
        type: string
      reservation_id:
        type: integer
      room:
        type: string
      seat:
        type: string
      seat_id:
        type: integer
      showtime_end:
        type: string
      showtime_id:
        type: integer
      showtime_start:
        type: string
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  main.ReservationStatus:
    enum:
    - unpaid
    - paid
    - cancelled
    type: string
    x-enum-varnames:
    - ReservationUnpaid
    - ReservationPaid
    - ReservationCancelled
  main.ReservationUserCreateReq:
    properties:
      cart_ids:
        items:
          type: integer
        type: array
    type: object
  main.Response-any:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  main.Response-main_Paginate-main_Reservation:
    properties:
      data:
        $ref: '#/definitions/main.Paginate-main_Reservation'
      message:
        type: string
    type: object
...
  main.Response-main_Reservation:
    properties:
      data:
        $ref: '#/definitions/main.Reservation'
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/reservations:
    post:
      consumes:
      - application/json
      description: user create reservations
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ReservationUserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Create Reservation
      tags:
      - reservations
  /api/reservations/{id}:
    delete:
      consumes:
      - application/json
      description: user delete reservation by reservation by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: reservation id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Delete Reservation
      tags:
      - reservations
    get:
      consumes:
      - application/json
      description: user update reservation by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: reservation id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Update Reservation
      tags:
      - reservations
  /api/reservations/{id}/cancel:
    put:
      consumes:
      - application/json
      description: user cancel reservation by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: reservation id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Cancel Reservation
      tags:
      - reservations
  /api/reservations/{id}/pay:
    put:
      consumes:
      - application/json
      description: user pay reservation by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: reservation id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Pay Reservation
      tags:
      - reservations
  /api/reservations/filter:
    post:
      consumes:
      - application/json
      description: user filter own reservations
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: pagination page
        in: query
        name: page
        type: integer
      - description: pagination page size
        in: query
        name: per_page
        type: integer
      - description: filter
        in: body
        name: request
        schema:
          $ref: '#/definitions/main.ReservationFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Paginate-main_Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Filter Reservation
      tags:
      - reservations
...
swagger: "2.0"